Upgrade to v7.4 (#152)
* Upgrade to metalcloud-sdk-go v7.4

Adapt the CLI to the v7.4 SDK, which switched resource IDs and many model
fields from float32/int32 to int64 and reworked several APIs.

int64 migration:
- Use int64/PtrInt64/[]int64/NullableInt64 where SDK models now require it
  (storage, subnet, network_device, logical_network(_profile), endpoint,
  infrastructure, os_template, user, vm_instance, fabric, server_instance,
  bucket, drive, file_share, server_default_credentials).
- Cast int64 model IDs to float32 where API path params remain float32.
- Add utils.GetInt64FromString; fix extension.Revision formatting (%d).

API changes:
- Events: FilterSeverity -> FilterLevel.
- network_device: EnableNetworkDeviceSyslog now returns (job, res, err).
- firmware_catalog: vendor/updateType/severity are plain strings now;
  add firmware_catalog/constants.go with the allowed values.

Removed (no SDK equivalent):
- user limits update, user permissions get/update, infrastructure
  user-limits, and the firmware-policy command group.

Reworked:
- user limits get -> UsersAPI.GetQuotaLimitsBreakdown.
- Startup command-gating is now permissive (the SDK no longer exposes a
  per-user permission list); the server still enforces authorization.

New command groups (SecurityAPI):
- quota-profile (CRUD) and permission create/delete/config-example.

+++ internal/drive/drive_snapshot.go
@@ -42,7 +42,7 @@ func DriveSnapshotList(ctx context.Context, infrastructureIdOrLabel string, driv
 	client := api.GetApiClient(ctx)
 
 	// GetDriveSnapshots returns a flat []SharedDriveSnapshot — no Page/Limit methods; not paginated.
-	snapshots, httpRes, err := client.DriveAPI.GetDriveSnapshots(ctx, infrastructureInfo.Id, driveIdNumeric).Execute()
+	snapshots, httpRes, err := client.DriveAPI.GetDriveSnapshots(ctx, int64(infrastructureInfo.Id), driveIdNumeric).Execute()
 	if err := response_inspector.InspectResponse(httpRes, err); err != nil {
 		return err
 	}
@@ -65,7 +65,7 @@ func DriveSnapshotCreate(ctx context.Context, infrastructureIdOrLabel string, dr
 
 	client := api.GetApiClient(ctx)
 
-	snapshot, httpRes, err := client.DriveAPI.CreateDriveSnapshot(ctx, infrastructureInfo.Id, driveIdNumeric).Execute()
+	snapshot, httpRes, err := client.DriveAPI.CreateDriveSnapshot(ctx, int64(infrastructureInfo.Id), driveIdNumeric).Execute()
 	if err := response_inspector.InspectResponse(httpRes, err); err != nil {
 		return err
 	}
@@ -90,7 +90,7 @@ func DriveSnapshotDelete(ctx context.Context, infrastructureIdOrLabel string, dr
 	client := api.GetApiClient(ctx)
 
 	httpRes, err := client.DriveAPI.
-		DeleteDriveSnapshot(ctx, infrastructureInfo.Id, driveIdNumeric).
+		DeleteDriveSnapshot(ctx, int64(infrastructureInfo.Id), driveIdNumeric).
 		DeleteSharedDriveSnapshot(sdk.DeleteSharedDriveSnapshot{
 			Name: snapshotName,
 		}).
@@ -119,7 +119,7 @@ func DriveSnapshotRestore(ctx context.Context, infrastructureIdOrLabel string, d
 	client := api.GetApiClient(ctx)
 
 	httpRes, err := client.DriveAPI.
-		RestoreDriveToSnapshot(ctx, infrastructureInfo.Id, driveIdNumeric).
+		RestoreDriveToSnapshot(ctx, int64(infrastructureInfo.Id), driveIdNumeric).
 		RestoreSharedDriveSnapshot(sdk.RestoreSharedDriveSnapshot{
 			Name: snapshotName,
 		}).
