+++ internal/file_share/file_share_snapshot.go
@@ -42,7 +42,7 @@ func FileShareSnapshotList(ctx context.Context, infrastructureIdOrLabel string, 
 	client := api.GetApiClient(ctx)
 
 	// GetFileShareSnapshots returns a flat []FileShareSnapshot — no Page/Limit methods; not paginated.
-	snapshots, httpRes, err := client.FileShareAPI.GetFileShareSnapshots(ctx, infrastructureInfo.Id, fileShareIdNumeric).Execute()
+	snapshots, httpRes, err := client.FileShareAPI.GetFileShareSnapshots(ctx, int64(infrastructureInfo.Id), fileShareIdNumeric).Execute()
 	if err := response_inspector.InspectResponse(httpRes, err); err != nil {
 		return err
 	}
@@ -65,7 +65,7 @@ func FileShareSnapshotCreate(ctx context.Context, infrastructureIdOrLabel string
 
 	client := api.GetApiClient(ctx)
 
-	snapshot, httpRes, err := client.FileShareAPI.CreateFileShareSnapshot(ctx, infrastructureInfo.Id, fileShareIdNumeric).Execute()
+	snapshot, httpRes, err := client.FileShareAPI.CreateFileShareSnapshot(ctx, int64(infrastructureInfo.Id), fileShareIdNumeric).Execute()
 	if err := response_inspector.InspectResponse(httpRes, err); err != nil {
 		return err
 	}
@@ -90,7 +90,7 @@ func FileShareSnapshotDelete(ctx context.Context, infrastructureIdOrLabel string
 	client := api.GetApiClient(ctx)
 
 	httpRes, err := client.FileShareAPI.
-		DeleteFileShareSnapshot(ctx, infrastructureInfo.Id, fileShareIdNumeric).
+		DeleteFileShareSnapshot(ctx, int64(infrastructureInfo.Id), fileShareIdNumeric).
 		DeleteFileShareSnapshot(sdk.DeleteFileShareSnapshot{
 			Name: snapshotName,
 		}).
@@ -119,7 +119,7 @@ func FileShareSnapshotRestore(ctx context.Context, infrastructureIdOrLabel strin
 	client := api.GetApiClient(ctx)
 
 	httpRes, err := client.FileShareAPI.
-		RestoreFileShareToSnapshot(ctx, infrastructureInfo.Id, fileShareIdNumeric).
+		RestoreFileShareToSnapshot(ctx, int64(infrastructureInfo.Id), fileShareIdNumeric).
 		RestoreFileShareSnapshot(sdk.RestoreFileShareSnapshot{
 			Name: snapshotName,
 		}).
